Folder restructure into components of optimisation
No change to the code other than file locations. Aiming to represent the components of the architecture diagram in the structure of the code.
