--- pybop/identification/__init__.py
@@ -1,4 +1,0 @@
-from .parameter_set import ParameterSet
-from .parameter import Parameter
-from .observations import Observed
-from .parameterisation import Parameterisation
