+++ pybop/optimisers/SciPyMinimize.py
